Follow symlinks
If a path assumed to point at a directory to be searched happens to be
a symlink, follow it.

The NULL check is hoisted because g_file_test assumes filename !=
NULL.

Also add an existence check for the directory argument to
`add_dir_to_watch` which fails silently unless path exists, imitating
past behaviour and accomodating following check for symlink-ness,
i.e. unless this is checked first there will be spurious error reports
for paths not found as the symlink is followed.